% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsFourWay-house-keeping.R
\name{four.col.counter}
\alias{four.col.counter}
\title{-------------------
Counting functions
-------------------}
\usage{
four.col.counter(dat, lfc)
}
\description{
-------------------
Counting functions
-------------------
}
