% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsFourWay-house-keeping.R
\name{four.comp1}
\alias{four.comp1}
\title{----------------------------------
Four Way plot component functions
----------------------------------}
\usage{
four.comp1(x.lim, y.lim, padj, all.lab, lfc, b, g, r)
}
\description{
----------------------------------
Four Way plot component functions
----------------------------------
}
