% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsFourWay-house-keeping.R
\name{four.out.ranker}
\alias{four.out.ranker}
\title{------------------
Ranking functions
------------------}
\usage{
four.out.ranker(px, py, x.lim, y.lim)
}
\description{
------------------
Ranking functions
------------------
}
