% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsScatterMatrix-house-keeping.R
\name{getCuffScatterMatrix}
\alias{getCuffScatterMatrix}
\title{--------------------------------------------------------#
Title:  ggviseq - House Keeping - scatterplot matrix   #
Author: Brandon Monier (brandon.monier@sdstate.edu)    #
Date:   04.10.17                                       #
--------------------------------------------------------#
---------
Preamble
---------
...
------------------------------
Scatterplot matrix extraction
------------------------------
cuffdiff}
\usage{
getCuffScatterMatrix(data)
}
\description{
--------------------------------------------------------#
Title:  ggviseq - House Keeping - scatterplot matrix   #
Author: Brandon Monier (brandon.monier@sdstate.edu)    #
Date:   04.10.17                                       #
--------------------------------------------------------#
---------
Preamble
---------
...
------------------------------
Scatterplot matrix extraction
------------------------------
cuffdiff
}
