% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcano-house-keeping.R,
%   R/vsVolcanoMatrix-house-keeping.R
\name{getDeseqVolcano}
\alias{getDeseqVolcano}
\alias{getDeseqVolcano}
\title{DESeq2 - NO PREREQUISITES}
\usage{
getDeseqVolcano(x, y, data, d.factor)

getDeseqVolcano(x, y, data, d.factor)
}
\description{
DESeq2 - NO PREREQUISITES

DESeq2 - NO PREREQUISITES
}
