% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsBoxPlot-house-keeping.R
\name{getEdgeBox}
\alias{getEdgeBox}
\title{-----------------------------------------------------#
Title:  ggviseq - Data Extractors - Box plots       #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
--------------------
Box plot extraction
--------------------
edgeR}
\usage{
getEdgeBox(data)
}
\description{
-----------------------------------------------------#
Title:  ggviseq - Data Extractors - Box plots       #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
--------------------
Box plot extraction
--------------------
edgeR
}
