% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsDEGmat-house-keeping.R
\name{getEdgeDEGMat}
\alias{getEdgeDEGMat}
\title{-----------------------------------------------------#
Title:  ggviseq - House Keeping - DEG Matrix        #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
---------------------------
DEG Matrix plot extraction
---------------------------
edgeR}
\usage{
getEdgeDEGMat(data, padj)
}
\description{
-----------------------------------------------------#
Title:  ggviseq - House Keeping - DEG Matrix        #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
---------------------------
DEG Matrix plot extraction
---------------------------
edgeR
}
