% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsFourWay-house-keeping.R
\name{getEdgeFourWay}
\alias{getEdgeFourWay}
\title{-----------------------------------------------------#
Title:  ggviseq - House Keeping - four way plot     #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   03.28.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
-------------------------
Four Way plot extraction
-------------------------
edgeR - REQUIRES `getEdgeScatter()`}
\usage{
getEdgeFourWay(x, y, control, data)
}
\description{
-----------------------------------------------------#
Title:  ggviseq - House Keeping - four way plot     #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   03.28.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
-------------------------
Four Way plot extraction
-------------------------
edgeR - REQUIRES `getEdgeScatter()`
}
