% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAPlot-house-keeping.R
\name{getEdgeMA}
\alias{getEdgeMA}
\title{-----------------------------------------------------#
Title:  ggviseq - House Keeping - MA plots          #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
-------------------
MA plot extraction
-------------------
edgeR - REQUIRES `getEdgeScatter()`}
\usage{
getEdgeMA(x, y, data)
}
\description{
-----------------------------------------------------#
Title:  ggviseq - House Keeping - MA plots          #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
-------------------
MA plot extraction
-------------------
edgeR - REQUIRES `getEdgeScatter()`
}
