% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsScatterPlot-house-keeping.R
\name{getEdgeScatter}
\alias{getEdgeScatter}
\title{-----------------------------------------------------#
Title:  ggviseq - House Keeping - scatter plots     #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
------------------------
Scatter plot extraction
------------------------
edgeR}
\usage{
getEdgeScatter(x, y, data)
}
\description{
-----------------------------------------------------#
Title:  ggviseq - House Keeping - scatter plots     #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
------------------------
Scatter plot extraction
------------------------
edgeR
}
