% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcano-house-keeping.R,
%   R/vsVolcanoMatrix-house-keeping.R
\name{getEdgeVolcano}
\alias{getEdgeVolcano}
\alias{getEdgeVolcano}
\title{------------------------
Volcano plot extraction
------------------------
edgeR - REQUIRES `getEdgeScatter()`}
\usage{
getEdgeVolcano(x, y, data)

getEdgeVolcano(x, y, data)
}
\description{
------------------------
Volcano plot extraction
------------------------
edgeR - REQUIRES `getEdgeScatter()`

------------------------
Volcano plot extraction
------------------------
edgeR - REQUIRES `getEdgeScatter()`
}
