% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAPlot-house-keeping.R
\name{ma.col.counter}
\alias{ma.col.counter}
\title{-------------------
Counting functions
-------------------}
\usage{
ma.col.counter(dat, lfc)
}
\description{
-------------------
Counting functions
-------------------
}
