% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAPlot-house-keeping.R
\name{ma.comp1}
\alias{ma.comp1}
\title{----------------------------
MA plot component functions
----------------------------}
\usage{
ma.comp1(y.lim, padj, lfc, b, g)
}
\description{
----------------------------
MA plot component functions
----------------------------
}
