% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAPlot-house-keeping.R
\name{ma.out.ranker}
\alias{ma.out.ranker}
\title{------------------
Ranking functions
------------------}
\usage{
ma.out.ranker(log2fc, lim)
}
\description{
------------------
Ranking functions
------------------
}
