% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAPlot-house-keeping.R
\name{ma.ranker}
\alias{ma.ranker}
\title{For data return only (optimize at later date)}
\usage{
ma.ranker(data, padj, lfc, y.lim)
}
\description{
For data return only (optimize at later date)
}
