% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcano-house-keeping.R
\name{vo.col.counter}
\alias{vo.col.counter}
\title{-------------------
Counting functions
-------------------}
\usage{
vo.col.counter(dat, lfc)
}
\description{
-------------------
Counting functions
-------------------
}
