% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcano-house-keeping.R
\name{vo.comp1}
\alias{vo.comp1}
\title{-----------------------------------------------------#
Title:  ggviseq - Data Extractors - Volcano Plots   #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
---------------------------------
Volcano plot component functions
---------------------------------}
\usage{
vo.comp1(x.lim, padj, lfc, b, g)
}
\description{
-----------------------------------------------------#
Title:  ggviseq - Data Extractors - Volcano Plots   #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.04.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
---------------------------------
Volcano plot component functions
---------------------------------
}
