% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcano-house-keeping.R
\name{vo.out.ranker}
\alias{vo.out.ranker}
\title{------------------
Ranking functions
------------------}
\usage{
vo.out.ranker(log2fc, lim)
}
\description{
------------------
Ranking functions
------------------
}
