% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcano-house-keeping.R
\name{vo.ranker}
\alias{vo.ranker}
\title{For data return only (optimize at later date)}
\usage{
vo.ranker(data, padj, lfc, x.lim)
}
\description{
For data return only (optimize at later date)
}
