% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcanoMatrix-house-keeping.R
\name{vomat.comp}
\alias{vomat.comp}
\title{-----------------------------------------------------#
Title:  ggviseq - House Keeping - volcano matrix    #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.27.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
--------------------------
Volcano matrix components
--------------------------}
\usage{
vomat.comp(padj, lfc)
}
\description{
-----------------------------------------------------#
Title:  ggviseq - House Keeping - volcano matrix    #
Author: Brandon Monier (brandon.monier@sdstate.edu) #
Date:   04.27.17                                    #
-----------------------------------------------------#
---------
Preamble
---------
...
--------------------------
Volcano matrix components
--------------------------
}
