% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsBoxPlot.R
\name{vsBoxPlot}
\alias{vsBoxPlot}
\title{Box plot for log10(FPKM or CPM) distributions}
\usage{
vsBoxPlot(data, d.factor = NULL, type, title = TRUE, legend = TRUE,
  grid = TRUE)
}
\arguments{
\item{data}{a cuffdiff, DESeq2, or edgeR object.}

\item{d.factor}{a specified factor; for use with DESeq2 objects only. Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the data. Must be either `cuffdiff`, `deseq`, or `edgeR`.}

\item{title}{show title of plot. Defaults to `TRUE`.}

\item{legend}{shows legend of plot. Defaults to `TRUE`.}

\item{grid}{show major and minor axis lines. Defaults to `TRUE`.}

\item{man.title}{a manually specified title at the authors discretion. Defaults to `NULL`.}
}
\description{
This function allows you to extract necessary results-based data from analytical objects
to create a box plot comparing log10(FPKM or CPM) distributions for experimental treatments.
}
\examples{
# Cuffdiff example
load('df.cuff.RData')
vsScatterMatrix(data = df.cuff, d.factor = NULL, type = 'cuffdiff', 
                title = TRUE, grid = TRUE)

# DESeq2 example
load('df.deseq.RData')
vsScatterMatrix(data = df.deseq, d.factor = 'cell', type = 'deseq',
                title = TRUE, grid = FALSE)

# edgeR example
load('df.edger.RData')
vsScatterMatrix(data = df.edger, d.factor = NULL, type = 'edger', 
                title = TRUE, grid = TRUE)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
