% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsDEGmat.R
\name{vsDEGMatrix}
\alias{vsDEGMatrix}
\title{Differential gene expression matrix}
\usage{
vsDEGMatrix(data, padj = 0.05, d.factor = NULL, type = c("cuffdiff",
  "deseq", "edger"), title = TRUE, legend = TRUE, grid = TRUE)
}
\arguments{
\item{data}{output generated from calling the main routines of either
`cuffdiff`, `DESeq2`, or `edgeR` analyses. For `cuffdiff`, this will be a 
`*_exp.diff` file. For `DESeq2`, this will be a generated object of class 
`DESeqDataSet`. For `edgeR`, this will be a generated object of class 
`DGEList`.}

\item{padj}{a user defined adjusted p-value cutoff point. Defaults to 
`0.05`.}

\item{d.factor}{a specified factor; for use with `DESeq2` objects only.
This input equates to the first parameter for the contrast argument when 
invoking the `results()` function in `DESeq2`. Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edger`. `cuffdiff` must be 
used with `cuffdiff` data; `deseq` must be used for `DESeq2` output; 
`edgeR` must be used with `edgeR` data. See the `data` parameter for 
further details.}

\item{title}{display the main title of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no title will display in plot.}

\item{legend}{display legend of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no legend will display in plot.}

\item{grid}{display major and minor axis lines. Logical; defaults to `TRUE`.
If set to `FALSE`, no axis lines will display in plot.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to visualize the number of differentially 
 expressed genes 
(DEG) at a given #' adjusted P-value for each experimental treatment level.
 Higher color intensity correlates to a higher number of DEGs.
}
\examples{
# cuffdiff example
data("df.cuff")
vsDEGMatrix(
 df.cuff, padj = 0.05, d.factor = NULL, type = 'cuffdiff', 
 title = TRUE, legend = TRUE, grid = TRUE
)

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsDEGMatrix(
 df.deseq, padj = 0.05, d.factor = 'condition', type = 'deseq', 
 title = TRUE, legend = TRUE, grid = TRUE
)

# edgeR example
data("df.edger")
require(edgeR)
vsDEGMatrix(
 df.edger, padj = 0.05, d.factor = NULL, type = 'edger', 
 title = TRUE, legend = TRUE, grid = TRUE
)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
