% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsBoxPlot.R
\name{vsBoxPlot}
\alias{vsBoxPlot}
\title{Box plot for log10(FPKM or CPM) distributions}
\usage{
vsBoxPlot(data, d.factor = NULL, type = c("cuffdiff", "deseq", "edger"),
  title = TRUE, legend = TRUE, grid = TRUE)
}
\arguments{
\item{data}{output generated from calling the main routines of either
`cuffdiff`, `DESeq2`, or `edgeR` analyses. For `cuffdiff`, this will be a 
`*_exp.diff` file. For `DESeq2`, this will be a generated object of class 
`DESeqDataSet`. For `edgeR`, this will be a generated object of class 
`DGEList`.}

\item{d.factor}{a specified factor; for use with `DESeq2` objects only.
This input equates to the first parameter for the contrast argument when 
invoking the `results()` function in `DESeq2`. Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edger`. `cuffdiff` must be 
used with `cuffdiff` data; `deseq` must be used for `DESeq2` output; 
`edgeR` must be used with `edgeR` data. See the `data` parameter for 
further details.}

\item{title}{display the main title of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no title will display in plot.}

\item{legend}{display legend of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no legend will display in plot.}

\item{grid}{display major and minor axis lines. Logical; defaults to `TRUE`.
If set to `FALSE`, no axis lines will display in plot.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to extract necessary results-based data from
 analytical objects
to create a box plot comparing log10(FPKM or CPM) distributions for 
 experimental treatments.
}
\examples{
# Cuffdiff example
data("df.cuff")
vsBoxPlot(
 data = df.cuff, d.factor = NULL, type = 'cuffdiff', title = TRUE,
 legend = TRUE, grid = TRUE
)

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsBoxPlot(
 data = df.deseq, d.factor = 'condition', type = 'deseq', 
 title = TRUE, legend = TRUE, grid = TRUE
)

# edgeR example
data("df.deseq")
require(edgeR)
vsBoxPlot(
 data = df.edger, d.factor = NULL, type = 'edger', title = TRUE,
 legend = TRUE, grid = TRUE
)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
