% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsDEGmat.R
\name{vsDEGMatrix}
\alias{vsDEGMatrix}
\title{Differential gene expression matrix}
\usage{
vsDEGMatrix(data, padj = 0.05, d.factor = NULL, type, title = TRUE,
  legend = TRUE, grid = TRUE)
}
\arguments{
\item{data}{a cuffdiff, DESeq2, or edgeR object.}

\item{padj}{a user defined adjusted p-value cutoff point. Defaults to 
`0.05`.}

\item{d.factor}{a specified factor; for use with DESeq2 objects only. 
Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the
data. Must be either `cuffdiff`, `deseq`, or `edgeR`.}

\item{title}{show title of plot. Defaults to `TRUE`.}

\item{legend}{show legend of plot. Defaults to `TRUE`.}

\item{grid}{show major and minor axis lines. Defaults to `TRUE`.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to visualize the number of differentially 
 expressed genes 
(DEG) at a given #' adjusted P-value for each experimental treatment level.
 Higher color intensity correlates to a higher number of DEGs.
}
\examples{
# cuffdiff example
data("df.cuff")
vsDEGMatrix(df.cuff, padj = 0.05, d.factor = NULL, type = 'cuffdiff', 
            title = TRUE, legend = TRUE, grid = TRUE)

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsDEGMatrix(df.deseq, padj = 0.05, d.factor = 'condition', type = 'deseq', 
            title = TRUE, legend = TRUE, grid = TRUE)

# edgeR example
data("df.edger")
require(edgeR)
vsDEGMatrix(df.edger, padj = 0.05, d.factor = NULL, type = 'edger', 
            title = TRUE, legend = TRUE, grid = TRUE)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
