% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAMatrix.R
\name{vsMAMatrix}
\alias{vsMAMatrix}
\title{MA plot matrix from log2 fold changes and -log10(p-values)}
\usage{
vsMAMatrix(data, d.factor = NULL, type, padj = 0.05, y.lim = NULL,
  lfc = NULL, title = TRUE, legend = TRUE, grid = TRUE, counts = TRUE,
  data.return = FALSE)
}
\arguments{
\item{data}{a cuffdiff, DESeq2, or edgeR object.}

\item{d.factor}{a specified factor; for use with DESeq2 objects only. 
Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edgeR`.}

\item{padj}{a user defined adjusted p-value cutoff point. 
Defaults to `0.05`.}

\item{y.lim}{set manual limits to the y axis. Defaults to `NULL`.}

\item{lfc}{log fold change level for setting conditonals. If no user input 
is added (`NULL`), value defaults to `1`.}

\item{title}{show title of plot. Defaults to `TRUE`.}

\item{legend}{shows legend of plot. Defaults to `TRUE`.}

\item{grid}{show major and minor axis lines. Defaults to `TRUE`.}

\item{counts}{displays the number of differentially expressed genes for 
each treatment comparison. Defaults to `TRUE`.}

\item{data.return}{returns data output of plot if set to `TRUE`. Defaults 
to `FASLSE`.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to generate MA plots for all possible treatment
combinations for a given factor in either a cuffdiff, DESeq2, or edgeR
data set.
}
\examples{
# Cuffdiff example
data("df.cuff")
vsMAMatrix(
 data = df.cuff, d.factor = NULL, type = 'cuffdiff', 
 padj = 0.05, y.lim = NULL, lfc = 1, title = TRUE, 
 grid = TRUE, counts = TRUE, data.return = FALSE
)

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsMAMatrix(
 data = df.deseq, d.factor = 'condition', type = 'deseq', 
 padj = 0.05, y.lim = NULL, lfc = 1, title = TRUE, 
 grid = TRUE, counts = TRUE, data.return = FALSE
)

# edgeR example
data("df.edger")
require(edgeR)
vsMAMatrix(
 data = df.edger, d.factor = NULL, type = 'edger', 
 padj = 0.05, y.lim = NULL, lfc = 1, title = TRUE, 
 grid = TRUE, counts = TRUE, data.return = FALSE
)
                
# Extract data frame from visualization
data("df.cuff")
tmp <- vsMAMatrix(
 data = df.cuff, d.factor = NULL, 
 type = 'cuffdiff', padj = 0.05, y.lim = NULL,
 lfc = 1, title = TRUE, grid = TRUE, 
 counts = TRUE, data.return = TRUE
)
df.vmat <- tmp[[1]]
head(df.vmat)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
