% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAMatrix.R
\name{vsMAMatrix}
\alias{vsMAMatrix}
\title{MA plot matrix from log2 fold changes and -log10(p-values)}
\usage{
vsMAMatrix(data, d.factor = NULL, type = c("cuffdiff", "deseq", "edger"),
  padj = 0.05, y.lim = NULL, lfc = NULL, title = TRUE, legend = TRUE,
  grid = TRUE, counts = TRUE, data.return = FALSE)
}
\arguments{
\item{data}{output generated from calling the main routines of either
`cuffdiff`, `DESeq2`, or `edgeR` analyses. For `cuffdiff`, this will be a 
`*_exp.diff` file. For `DESeq2`, this will be a generated object of class 
`DESeqDataSet`. For `edgeR`, this will be a generated object of class 
`DGEList`.}

\item{d.factor}{a specified factor; for use with `DESeq2` objects only.
This input equates to the first parameter for the contrast argument when 
invoking the `results()` function in `DESeq2`. Defaults to `NULL`.}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edger`. `cuffdiff` must be 
used with `cuffdiff` data; `deseq` must be used for `DESeq2` output; 
`edgeR` must be used with `edgeR` data. See the `data` parameter for 
further details.}

\item{padj}{a user defined adjusted p-value cutoff point. 
Defaults to `0.05`.}

\item{y.lim}{set manual limits (boundaries) to the y axis. Defaults to 
`NULL`.}

\item{lfc}{log fold change level for setting conditonals. If no user input 
is added (`NULL`), value defaults to `1`.}

\item{title}{display the main title of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no title will display in plot.}

\item{legend}{display legend of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no legend will display in plot.}

\item{grid}{display major and minor axis lines. Logical; defaults to `TRUE`.
If set to `FALSE`, no axis lines will display in plot.}

\item{counts}{displays the number of differentially expressed genes for 
each treatment comparison. Defaults to `TRUE`.}

\item{data.return}{returns data output of plot Logical; defaults to `FALSE`.
If set to `TRUE`, a data frame will also be called. Assign to object
for reproduction and saving of data frame. See final example for further
details.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to generate MA plots for all possible treatment
combinations for a given factor in either a cuffdiff, DESeq2, or edgeR
data set.
}
\examples{
# Cuffdiff example
data("df.cuff")
vsMAMatrix(
 data = df.cuff, d.factor = NULL, type = 'cuffdiff', 
 padj = 0.05, y.lim = NULL, lfc = 1, title = TRUE, 
 grid = TRUE, counts = TRUE, data.return = FALSE
)

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsMAMatrix(
 data = df.deseq, d.factor = 'condition', type = 'deseq', 
 padj = 0.05, y.lim = NULL, lfc = 1, title = TRUE, 
 grid = TRUE, counts = TRUE, data.return = FALSE
)

# edgeR example
data("df.edger")
require(edgeR)
vsMAMatrix(
 data = df.edger, d.factor = NULL, type = 'edger', 
 padj = 0.05, y.lim = NULL, lfc = 1, title = TRUE, 
 grid = TRUE, counts = TRUE, data.return = FALSE
)
                
# Extract data frame from visualization
data("df.cuff")
tmp <- vsMAMatrix(
 data = df.cuff, d.factor = NULL, 
 type = 'cuffdiff', padj = 0.05, y.lim = NULL,
 lfc = 1, title = TRUE, grid = TRUE, 
 counts = TRUE, data.return = TRUE
)
df.vmat <- tmp[[1]]
head(df.vmat)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
