% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAPlot.R
\name{vsMAPlot}
\alias{vsMAPlot}
\title{MA plot from mean expression and log fold changes from different analytical 
objects}
\usage{
vsMAPlot(x, y, data, d.factor = NULL, type, padj = 0.1, y.lim = NULL,
  lfc = NULL, title = TRUE, legend = TRUE, grid = TRUE,
  data.return = FALSE)
}
\arguments{
\item{x}{treatment `x` for comparison (log2(y/x)).}

\item{y}{treatment `y` for comparison (log2(y/x)).}

\item{data}{a cuffdiff, DESeq2, or edgeR object.}

\item{d.factor}{a specified factor; for use with DESeq2 objects only.
Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edgeR`.}

\item{padj}{a user defined adjusted p-value cutoff point. Defaults to `0.1`.}

\item{y.lim}{set manual limits to the y axis. Defaults to `NULL`.}

\item{lfc}{log fold change level for setting conditonals. If no user input 
is added (`NULL`), value defaults to `1`.}

\item{title}{show title of plot. Defaults to `TRUE`.}

\item{legend}{shows legend of plot. Defaults to `TRUE`.}

\item{grid}{show major and minor axis lines. Defaults to `TRUE`.}

\item{data.return}{returns data output of plot if set to `TRUE`. Defaults 
to `FASLSE`.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to extract necessary results-based data from 
different output files to create a MA plot (i.e. a scatter plot) of log2 
fold changes versus normalized mean counts while implementing ggplot2 
aesthetics.
}
\examples{
# Cuffdiff example
data("df.cuff")
vsMAPlot(x = 'hESC', y = 'iPS', data = df.cuff, d.factor = NULL, 
         type = 'cuffdiff', padj = 0.05, y.lim = NULL, lfc = 2, 
         title = TRUE, legend = TRUE, grid = TRUE, data.return = FALSE)

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsMAPlot(x = 'treated', y = 'untreated', data = df.deseq, 
         d.factor = 'condition', type = 'deseq', padj = 0.05, 
         y.lim = NULL, lfc = NULL, title = TRUE, legend = TRUE, 
         grid = TRUE, data.return = FALSE)

# edgeR example
data("df.edger")
require(edgeR)
vsMAPlot(x = 'WM', y = 'MM', data = df.edger, d.factor = NULL, 
         type = 'edger', padj = 0.1, y.lim = NULL, lfc = 2, 
         title = FALSE, legend = TRUE, grid = TRUE, data.return = FALSE)
                
# Extract data frame from visualization
data("df.cuff")
tmp <- vsMAPlot(x = 'hESC', y = 'iPS', data = df.cuff, 
                d.factor = NULL, type = 'cuffdiff', padj = 0.05, 
                y.lim = NULL, lfc = 2, title = TRUE, grid = TRUE, 
                data.return = TRUE)
df.ma <- tmp[[1]]
head(df.ma)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
