% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsMAPlot.R
\name{vsMAPlot}
\alias{vsMAPlot}
\title{MA plot from mean expression and log fold changes from different analytical 
objects}
\usage{
vsMAPlot(x, y, data, d.factor = NULL, type = c("cuffdiff", "deseq",
  "edger"), padj = 0.05, y.lim = NULL, lfc = NULL, title = TRUE,
  legend = TRUE, grid = TRUE, data.return = FALSE)
}
\arguments{
\item{x}{treatment `x` for comparison (log2(x/control)). This will be a
factor level in your data.}

\item{y}{treatment `y` for comparison (log2(y/control)). This will be a 
factor level in your data.}

\item{data}{output generated from calling the main routines of either
`cuffdiff`, `DESeq2`, or `edgeR` analyses. For `cuffdiff`, this will be a 
`*_exp.diff` file. For `DESeq2`, this will be a generated object of class 
`DESeqDataSet`. For `edgeR`, this will be a generated object of class 
`DGEList`.}

\item{d.factor}{a specified factor; for use with `DESeq2` objects only.
This input equates to the first parameter for the contrast argument when 
invoking the `results()` function in `DESeq2`. Defaults to `NULL`.}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edger`. `cuffdiff` must be 
used with `cuffdiff` data; `deseq` must be used for `DESeq2` output; 
`edgeR` must be used with `edgeR` data. See the `data` parameter for 
further details.}

\item{padj}{a user defined adjusted p-value cutoff point. 
Defaults to `0.05`.}

\item{y.lim}{set manual limits (boundaries) to the y axis. Defaults to 
`NULL`.}

\item{lfc}{log fold change level for setting conditonals. If no user input 
is added (`NULL`), value defaults to `1`.}

\item{title}{display the main title of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no title will display in plot.}

\item{legend}{display legend of plot. Logical; defaults to `TRUE`.
If set to `FALSE`, no legend will display in plot.}

\item{grid}{display major and minor axis lines. Logical; defaults to `TRUE`.
If set to `FALSE`, no axis lines will display in plot.}

\item{data.return}{returns data output of plot Logical; defaults to `FALSE`.
If set to `TRUE`, a data frame will also be called. Assign to object
for reproduction and saving of data frame. See final example for further
details.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to extract necessary results-based data from 
different output files to create a MA plot (i.e. a scatter plot) of log2 
fold changes versus normalized mean counts while implementing ggplot2 
aesthetics.
}
\examples{
# Cuffdiff example
data("df.cuff")
vsMAPlot(
 x = 'hESC', y = 'iPS', data = df.cuff, d.factor = NULL, 
 type = 'cuffdiff', padj = 0.05, y.lim = NULL, lfc = 1, 
 title = TRUE, legend = TRUE, grid = TRUE, data.return = FALSE
)

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsMAPlot(
 x = 'treated_paired.end', y = 'untreated_paired.end', 
 data = df.deseq, d.factor = 'condition', type = 'deseq', 
 padj = 0.05, y.lim = NULL, lfc = NULL, title = TRUE, 
 legend = TRUE, grid = TRUE, data.return = FALSE
)

# edgeR example
data("df.edger")
require(edgeR)
vsMAPlot(
 x = 'WM', y = 'MM', data = df.edger, d.factor = NULL, 
 type = 'edger', padj = 0.1, y.lim = NULL, lfc = 1, 
 title = FALSE, legend = TRUE, grid = TRUE, data.return = FALSE
)
                
# Extract data frame from visualization
data("df.cuff")
tmp <- vsMAPlot(
 x = 'hESC', y = 'iPS', data = df.cuff, 
 d.factor = NULL, type = 'cuffdiff', padj = 0.05, 
 y.lim = NULL, lfc = 1, title = TRUE, grid = TRUE, 
 data.return = TRUE
)
df.ma <- tmp[[1]]
head(df.ma)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
