% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsScatterMatrix.R
\name{vsScatterMatrix}
\alias{vsScatterMatrix}
\title{Scatter plot matrix of log10(FPKM or CPM) values}
\usage{
vsScatterMatrix(data, d.factor = NULL, type, comp = NULL, title = TRUE,
  grid = TRUE, man.title = NULL)
}
\arguments{
\item{data}{a cuffdiff, DESeq2, or edgeR object.}

\item{d.factor}{a specified factor; for use with DESeq2 objects only. 
Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edgeR`.}

\item{comp}{treatments you would like to compare in the form of a vector. 
If no parameter is specified, all possible treatment comparisons will be 
made. Defaults to `NULL`.}

\item{title}{show title of plot. Defaults to `TRUE`.}

\item{grid}{show major and minor axis lines. Defaults to `TRUE`.}

\item{man.title}{a manually specified title at the authors discretion. 
Defaults to `NULL`.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function will generate a matrix of scatterplots for all possible
treatment combinations with additional distribution info.
}
\examples{
# Cuffdiff example
data("df.cuff")
vsScatterMatrix(data = df.cuff, d.factor = NULL, type = 'cuffdiff', 
                comp = NULL, title = TRUE, grid = TRUE, 
                man.title = 'Example title')

# DESeq2 example
data("df.deseq")
require(DESeq2)
vsScatterMatrix(data = df.deseq, d.factor = 'condition', type = 'deseq',
                comp = NULL, title = TRUE, grid = FALSE, man.title = NULL)

# edgeR example
data("df.edger")
require(edgeR)
vsScatterMatrix(data = df.edger, d.factor = NULL, type = 'edger', 
                comp = c('WM', 'MM'), title = TRUE, grid = TRUE, 
                man.title = NULL)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
