% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsScatterPlot.R
\name{vsScatterPlot}
\alias{vsScatterPlot}
\title{Scatter plot of log10(FPKM or CPM) values}
\usage{
vsScatterPlot(x, y, data, d.factor = NULL, type, title = TRUE,
  grid = TRUE)
}
\arguments{
\item{x}{treatment for the x-axis of the plot}

\item{y}{treatment for the y-axis of the plot}

\item{data}{a cuffdiff, DESeq2, or edgeR object.}

\item{d.factor}{a specified factor; for use with DESeq2 objects only.
Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the 
data. Must be either `cuffdiff`, `deseq`, or `edgeR`.}

\item{title}{show title of plot. Defaults to `TRUE`.}

\item{grid}{show major and minor axis lines. Defaults to `TRUE`.}
}
\value{
An object created by \code{ggplot}
}
\description{
This function allows you to visualize comparisons of log10 values of either 
FPKM or CPM measurements of two treatments depending on analytical type.
}
\examples{
# Cuffdiff example
data("df.cuff")
vsScatterPlot(
 x = 'hESC', y = 'iPS', data = df.cuff, d.factor = NULL, 
 type = 'cuffdiff', title = TRUE, grid = TRUE
)

# DESeq2 example
data("df.deseq")
require(DESeq2)               
vsScatterPlot(
 x = 'treated_paired.end', y = 'untreated_paired.end', 
 data = df.deseq, d.factor = 'condition', type = 'deseq', 
 title = TRUE, grid = TRUE
)

# edgeR example
data("df.edger")
require(edgeR)
vsScatterPlot(
 x = 'WW', y = 'WM', data = df.edger, d.factor = NULL, 
 type = 'edger', title = TRUE, grid = TRUE
)
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
