% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsVolcanoMatrix.R
\name{vsVolcanoMatrix}
\alias{vsVolcanoMatrix}
\title{Volcano plot matrix from log2 fold changes and -log10(p-values)}
\usage{
vsVolcanoMatrix(data, d.factor = NULL, type, padj = 0.1, x.lim = NULL,
  lfc = NULL, title = TRUE, legend = TRUE, grid = TRUE, counts = TRUE,
  data.return = FALSE)
}
\arguments{
\item{data}{a cuffdiff, DESeq2, or edgeR object.}

\item{d.factor}{a specified factor; for use with DESeq2 objects only. Defaults to `NULL`}

\item{type}{an analysis classifier to tell the function how to process the data. Must be either `cuffdiff`, `deseq`, or `edgeR`.}

\item{padj}{a user defined adjusted p-value cutoff point. Defaults to `0.1`.}

\item{x.lim}{set manual limits to the x axis. Defaults to `NULL`.}

\item{lfc}{log fold change level for setting conditonals. If no user input is added (`NULL`), value defaults to `1`.}

\item{title}{show title of plot. Defaults to `TRUE`.}

\item{legend}{shows legend of plot. Defaults to `TRUE`.}

\item{grid}{show major and minor axis lines. Defaults to `TRUE`.}

\item{counts}{displays the number of differentially expressed genes for each treatment comparison. Defaults to `TRUE`.}

\item{data.return}{returns data output of plot if set to `TRUE`. Defaults to `FASLSE`.}
}
\description{
This function allows you to extract necessary results-based data from a 
DESEq object class to create a volcano plot (i.e. a scatter plot) of the
negative log of the p-value versus the log of the fold change while
implementing ggplot2 aesthetics for all possible combinations of treatments.
}
\examples{
# Cuffdiff example
load('df.cuff.RData')
vsVolcanoMatrix(data = df.cuff, d.factor = NULL, type = 'cuffdiff', 
                padj = 0.05, x.lim = NULL, lfc = 2, title = TRUE, 
                grid = TRUE, counts = TRUE, data.return = FALSE)

# DESeq2 example
load('df.deseq.RData')
vsVolcanoMatrix(data = df.deseq, d.factor = 'cell', type = 'deseq', 
                padj = 0.05, x.lim = NULL, lfc = 2, title = TRUE, 
                grid = TRUE, counts = TRUE, data.return = FALSE)

# edgeR example
load('df.edger.RData')
vsVolcanoMatrix(data = df.edger, d.factor = NULL, type = 'edger', 
                padj = 0.05, x.lim = NULL, lfc = 2, title = TRUE, 
                grid = TRUE, counts = TRUE, data.return = FALSE)
                
# Extract data frame from visualization
load('df.cuff.RData')
tmp <- vsVolcanoMatrix(x = 'hESC', y = 'iPS', data = df.cuff, 
                       d.factor = NULL, type = 'cuffdiff', padj = 0.05, 
                       x.lim = NULL, lfc = 2, title = TRUE, grid = TRUE, 
                       data.return = TRUE)
df.vmat <- tmp[[1]]
head(df.vmat)
Configure data
Conditional plot components
Configure data - aesthetic conditionals
Plot components
Conditonals - color counts
Plot
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
